requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
litellm==1.65.8
python-dotenv==1.1.0
pydantic==2.11.3
SQLAlchemy==2.0.40
python-multipart==0.0.20
